Fix fort we files (mostly)
* Converting WE files from one single line, to a line per node.
Also stripped the " " around "{", "}" and "=".

This will make the forthcoming changes easier to see in github differences.

Used the following commands (and regex) to accomplish this:
cd mods/lottmapgen/schems
vim gondorfort.we angmarfort.we rohanfort.we
:s/}, {/},\r{/g
:%s/ }/}/g
:%s/{ /{/g
:%s/ = /=/g

* Fix ladder positions ("param2") in angmarfort.we and gondorfort.we files.

Convert the static "default:chest" entries in rohanfort.we to "lottmapgen:rohan_chest_spawner". This is consistent with other buildings.

Also allow orc characters to use orc food and medicine without the blindness effect.