Lottfarming: Fix mushrooms spreading in weird stacks
The mushroom spread functions contained the code snippets for placing an
offspring twice, was this intentional?  I don't think so: if it was, why
was "if name==..." also repeated, instead of just calling set_node() twice.
And stacks of 2 mushrooms each look very weird, so I'm pretty sure this is
a bug.  Fixed the code.

Also simplified code for random position.